+++ CHANGELOG.md
@@ -6,6 +6,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### ADDED
 - Added support for the WideSky `createUser` endpoint, `/api/admin/user`. As its name implies, this function is use for 
   creating a new user account in WideSky through one of the supported local/scram authentication method.
+- Added support for the WideSky `watchSub` endpoint, `/api/watchSub`. This allows a user to subscribe to a watch.
+- Added a function to extend the lease of a watch.
+- Added support for the WideSky `watchUnsub` endpoint, `/api/watchUnsub`. This allows a user to unsubscribe to a watch.
+- Added a function to retrieve an API socket using a watch.
 
 ## [2.0.2] - 2022-11-07
 ### FIXED